chore: clean up tree logic
