+++ src/api/turing.ts
@@ -46,12 +46,9 @@ export const fetchTuringGame = async () => {
   for (let i = 1; i < moves.length; i++) {
     const move = moves[i]
     if (move.uci && move.san) {
-      currentNode = gameTree.addMainMove(
-        currentNode,
-        move.board,
-        move.uci,
-        move.san,
-      )
+      currentNode = gameTree
+        .getLastMainlineNode()
+        .addChild(move.board, move.uci, move.san, true)
     }
   }
